:construction: rename MobileDashboard to MobileDashboardView
